Attempt to extract geopyspark oriented job result metadata handling
and provide API for aggregator to inject alternative implementation

refs Open-EO/openeo-aggregator#204, #530, Open-EO/openeo-geopyspark-driver#440
