+++ openeo_driver/_version.py
@@ -1,1 +1,1 @@
-__version__ = "0.140.0a8"
+__version__ = "0.141.0a1"
